refactor cli helpers into generic env argument parser; rename examples
