+++ examples/09_interactive_envs/02_jumper/jumper_rule.py
@@ -1,67 +1,69 @@
 # SPDX-License-Identifier: MIT
-"""Run a simple rule-based steering controller on LineFollowerEnv."""
+"""Run a simple rule-based controller on JumperEnv."""
 
 import sys
 
 import pygame
 
-from evolib.evolib_envs.cli import parse_linefollower_args
+from evolib.evolib_envs.cli import parse_env_args
 from evolib.evolib_envs.core.env import Action, Observation
-from evolib.evolib_envs.envs.line_follower import LineFollowerEnv
-from evolib.evolib_envs.envs.line_follower_defaults import (
+from evolib.evolib_envs.envs.jumper import JumperEnv
+from evolib.evolib_envs.envs.jumper_defaults import (
     DEFAULT_FPS,
     DEFAULT_HEIGHT,
     DEFAULT_MAX_STEPS,
     DEFAULT_WIDTH,
 )
-from evolib.evolib_envs.renderers.pygame_line_follower import draw_env
+from evolib.evolib_envs.renderers.pygame_jumper import draw_env
 
 SCREEN_WIDTH = DEFAULT_WIDTH
 SCREEN_HEIGHT = DEFAULT_HEIGHT
 MAX_STEPS = DEFAULT_MAX_STEPS
 FPS = DEFAULT_FPS
 
-
-args = parse_linefollower_args()
+args = parse_env_args(description="Run a Jumper agent.")
 difficulty = args.difficulty
 
 
-class RuleBasedLineFollowerController:
-    """Simple rule-based steering controller using the two line sensors."""
+class RuleBasedJumperController:
+    """Jump when the obstacle is close and the player is on the ground."""
 
-    def __init__(self, *, turn_strength: float = 1.0) -> None:
-        self.turn_strength = turn_strength
+    def __init__(
+        self, *, jump_distance_min: float = 0.10, jump_distance_max: float = 0.22
+    ) -> None:
+        self.jump_distance_min = jump_distance_min
+        self.jump_distance_max = jump_distance_max
 
     def act(self, observation: Observation) -> Action:
-        left_sensor, right_sensor = observation
+        normalized_distance = observation[0]
 
-        error = right_sensor - left_sensor
-        turn = error * self.turn_strength
-        turn = max(-1.0, min(1.0, turn))
-
-        return [turn]
+        should_jump = (
+            self.jump_distance_min <= normalized_distance <= self.jump_distance_max
+        )
+        return [1.0, 0.75 if should_jump else 0.0, 0.0]
 
 
 def main() -> None:
-    env = LineFollowerEnv(
+    """Run the rule-based Jumper demo."""
+
+    env = JumperEnv(
         width=SCREEN_WIDTH,
         height=SCREEN_HEIGHT,
         max_steps=MAX_STEPS,
         difficulty=difficulty,
     )
-    controller = RuleBasedLineFollowerController()
+    controller = RuleBasedJumperController()
 
     observation = env.reset()
     total_reward = 0.0
 
     pygame.init()
     screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT))
-    pygame.display.set_caption("EvoLib Env - LineFollower Rule")
+    pygame.display.set_caption("EvoLib Env - Jumper Rule")
     clock = pygame.time.Clock()
     font = pygame.font.SysFont(None, 24)
 
     running = True
-
     while running:
         for event in pygame.event.get():
             if event.type == pygame.QUIT:
@@ -84,8 +86,7 @@ def main() -> None:
             observation = env.reset()
             total_reward = 0.0
 
-        draw_env(screen, env, total_reward, font, title="Rule-based LineFollower")
-
+        draw_env(screen, env, total_reward, font, title="Rule-based Jumper")
         pygame.display.flip()
         clock.tick(FPS)
